Update to include DAQS NIC configuration
Change to the buffer configuration tuning parameters governing
how network packet data is moved into the DAQ server memory for
processing.  The goal of this change is to reduce the time spent
by individual CPUs by spreading this load to more CPUs.

The default spreads this load to 8 of 32 CPUs, using 511 entry
packet in each of 8 buffers.  (4088 entries in total)

The change is to use 32 buffers with 2047 entries each.  (65504
entries in total)

At 250K samples per second, each chassis produces 25 MB/s in
~17500 packets per second.

Adaptive receive interrupt coalescence is enabled as part of this
tested configuration.

The change to set Channel configuration in 10-DAQeth0.link does not
have the documented effect as of systemd 252 circa Debian 12.
In the intrim, add manual setting change to the power up procedure